+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/
 spec/v2.0.0.html).
 
+## [1.25.19] 2026/03/05
+
+### Removed
+- Recurring tasks managed on the related goal show.
+
 ## [1.25.18] 2026/03/05
 
 ### Removed
-- Recurring task indicated on the show
+- Recurring task indicated on the show.
 
 ## [1.25.17] 2026/03/05